Add Authorization#verify_status
Recent versions of boulder will no longer process a challenge if the
associated authorization is already valid, that is another challenge was
previously solved. This means we need to allow people to poll on the
authorizations status rather than the challenge status so they don't
have to poll on the status of all challenges of an authorization all the
time.

See
https://community.letsencrypt.org/t/upcoming-change-valid-authz-reuse/16982
and https://github.com/letsencrypt/boulder/issues/2057