Implement Challenges::Base#[]
Acme::Client::Resources::Challenges.new may return a Hash instead of
Challenges::* when it encountered unsupported challenge type.

When an user is iterating authz#challenges, they have to check a class
of an element in prior to use.

Checking challenge#challenge_type is sometime useful than checking
challenge#class when an user has a support of additional challenge_type
which are uncovered in this gem.